pyproject.toml
[project]
name = "project-d"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django-phonenumber-field>=8.0.0",
    "django>=5.1.5",
    "requests>=2.32.3",
    "ruff>=0.9.3",
    "phonenumbers>=8.13.53",
    "django-crispy-forms>=2.3",
    "crispy-bootstrap5>=2024.10",
    "pylint-django>=2.6.1",
    "django-allauth>=65.3.1",
    "pyjwt>=2.10.1",
    "cryptography>=44.0.0",
    "black>=25.1.0",
    "git-filter-repo>=2.47.0",
    "psycopg2-binary>=2.9.10",
    "django-extensions>=3.2.3",
    "django-environ>=0.12.0",
    "gunicorn>=23.0.0",
    "pip>=25.0.1",
    "dj-database-url>=2.3.0",
    "whitenoise>=6.9.0",
    "psycopg>=3.2.6",
    "pillow>=11.1.0",
]
[tool.setuptools.packages.find]
include = ["config", "static", "test_app", "templates", "user_profile", "employee_learning"]
